Add files via upload
Pattern matching, a query operation based on target sequences, is the primary task of information retrieval. It is not only used to calculate the support or number of occurrences of a pattern in the field of knowledge mining but also has derived kinds of constraints during the search process. Among the variety of possible constraint-matching conditions, past researchers have considered the significant non-overlapping condition, which means that any two occurrences cannot use the identical character of the sequence in the same position of the pattern. Although a corresponding completeness proof has been given in state-of-the-art works, there is still room to improve the matching process under the non-overlapping constraint. Inspired by the literature on strict matching under the non-overlapping condition, this study proposes an approach to matching given patterns called NEONATE. Based on the Nettree index structure, we adopt a bi-directional search strategy and pruning operation to realize the non-overlapping pattern matching under the gap and length constraints. NEONATE first locates the horizontal layer with the fewest available nodes in the Nettree. Then we iteratively seek the rightmost parent and rightmost child node until an occurrence is obtained in which the nodes in the Nettree path can strictly satisfy the pattern constraint. We demonstrate the completeness of NEONATE and analyze the complexity of the algorithm. The correctness and efficiency of our approach are proven via extensive experiments.